[자동차 경주 Step5] refactor #time 1h30m 코드리뷰 반영  -1. RacingGame에서 ResultView 객체에게 view 로직 위임하도록 수정  -2. RaceResult 삭제 (view 로직 -> ResultView, 경주결과에 대한 로직 -> RacingGame)  -3. RacingGame.isFirstRace 삭제 (뷰에 대한 변수이므로 ResultView로 이동)
